Move OpenStack saving code under NetworkManager
Move OpenStack saving code under NetworkManager
